backend/requirements.txt
amqp==5.3.1
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
billiard==4.2.1
blinker==1.8.2
celery==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dep-logic==0.5.1
distlib==0.3.9
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
environ==1.0
exceptiongroup==1.3.0
filelock==3.16.1
findpython==0.6.3
h11==0.16.0
hishel==0.0.33
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
installer==0.7.0
kombu==5.5.3
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
npm==0.1.1
numpy==1.24.4
optional-django==0.1.0
packaging==25.0
pandas==2.0.3
pbs-installer==2025.4.9
pdm==2.20.1
pillow==10.4.0
pipdeptree==2.24.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
Pygments==2.19.1
PyJWT==1.7.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
resolvelib==1.1.0
rest-framework-simplejwt==0.0.2
rich==14.0.0
sentry-sdk==2.27.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sqlparse==0.5.1
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
unearth==0.17.5
urllib3==2.2.3
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
zipp==3.20.2
